+++ docs/source/examples/7/inputs.py
@@ -3,12 +3,13 @@
 #  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 #  ------------------------------------------------------------------------------------------
 from pathlib import Path
+from health_azure.utils import WORKSPACE_CONFIG_JSON
 
 import numpy as np
 from sklearn import datasets
 
-import health.azure.himl as himl
-from health.azure.datasets import get_datastore
+import health_azure.himl as himl
+from health_azure.datasets import get_datastore
 
 
 def main() -> None:
@@ -26,7 +27,7 @@ def main() -> None:
     path = Path(__file__).parent.resolve()
 
     workspace = himl.get_workspace(aml_workspace=None,
-                                   workspace_config_path=path / himl.WORKSPACE_CONFIG_JSON)
+                                   workspace_config_path=path / WORKSPACE_CONFIG_JSON)
 
     datastore = get_datastore(workspace=workspace,
                               datastore_name="himldatasets")