+++ docs/source/examples/7/sample.py
@@ -12,7 +12,7 @@ import numpy as np
 from sklearn.metrics import confusion_matrix
 from sklearn.model_selection import train_test_split
 
-from health.azure.himl import submit_to_azure_if_needed
+from health_azure.himl import submit_to_azure_if_needed
 
 
 def main() -> None:
@@ -30,8 +30,8 @@ def main() -> None:
                         help='Penalty parameter of the error term')
 
     args = parser.parse_args()
-    print('Kernel type:', np.str(args.kernel))
-    print('Penalty:', np.float(args.penalty))
+    print(f'Kernel type:{args.kernel}')
+    print(f'Penalty: {args.penalty}')
 
     # X -> features, y -> label
     input_folder = run_info.input_datasets[0] or Path("dataset")